Data extraction refactoring
GetData file was divided into two modules ('encoding' and 'decoding') that hold the I/O functions of the model, which are translations between MIDI and Multi Hot Encoding. This commit turns every m21 stram into a pandas dataframe that contains a header (with some info), metadata and/or a framestack containing the multi hot encoding of a measure indexed by note name and frame number.